server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
onnxruntime==1.18.1
ultralytics==8.2.47
streamlit==1.36.0
